requirements.txt
selenium
selenium-datatable
pytest
PyHamcrest